np.int is deprecated in numpy>=1.20 - use np.int32 instead

+++ pretty_midi/pretty_midi.py
@@ -321,7 +321,7 @@ class PrettyMIDI(object):
             last_note_on = collections.defaultdict(list)
             # Keep track of which instrument is playing in each channel
             # initialize to program 0 for all channels
-            current_instrument = np.zeros(16, dtype=np.int)
+            current_instrument = np.zeros(16, dtype=np.int32)
             for event in track:
                 # Look for track name events
                 if event.type == 'track_name':
